Moved Compose-related UI from identity-appsupport to multipaz-compose.
This is the first step toward organizing our projects into a better directory
structure. This is the only module that needs to be split into two before we
can start renaming and reorganizing all the others.

Tested-by:
- ./gradlew check
- ./gradlew connectedCheck
- Manually running the Wallet
- Manually running the testapp
