+++ multipaz-compose/build.gradle.kts
@@ -5,6 +5,8 @@ import org.jetbrains.kotlin.gradle.dsl.JvmTarget
 plugins {
     alias(libs.plugins.kotlinMultiplatform)
     alias(libs.plugins.androidLibrary)
+    alias(libs.plugins.jetbrainsCompose)
+    alias(libs.plugins.compose.compiler)
 }
 
 kotlin {
@@ -38,32 +40,35 @@ kotlin {
     sourceSets {
         val commonMain by getting {
             dependencies {
+                implementation(compose.runtime)
+                implementation(compose.foundation)
+                implementation(compose.material3)
+                implementation(compose.ui)
+                implementation(compose.components.resources)
+                implementation(compose.components.uiToolingPreview)
+                implementation(compose.materialIconsExtended)
+                implementation(libs.jetbrains.navigation.compose)
+                implementation(libs.jetbrains.navigation.runtime)
+
                 implementation(project(":identity"))
                 implementation(project(":identity-mdoc"))
-                implementation(libs.kotlinx.coroutines.core)
+                implementation(project(":identity-appsupport"))
                 implementation(libs.kotlinx.datetime)
-                implementation(libs.kotlinx.io.core)
-                implementation(libs.kotlinx.serialization.json)
                 implementation(libs.qrose)
                 implementation(libs.easyqrscan)
             }
         }
-
         val androidMain by getting {
             dependencies {
-                implementation(libs.bouncy.castle.bcprov)
-                implementation(libs.bouncy.castle.bcpkix)
-                implementation(libs.tink)
                 implementation(libs.accompanist.permissions)
                 implementation(libs.androidx.material)
-                implementation(libs.play.services.identity.credentials)
             }
         }
     }
 }
 
 android {
-    namespace = "com.android.identity.appsupport"
+    namespace = "org.multipaz.compose"
     compileSdk = libs.versions.android.compileSdk.get().toInt()
 
     defaultConfig {
@@ -76,6 +81,8 @@ android {
         targetCompatibility = JavaVersion.VERSION_17
     }
     dependencies {
+        debugImplementation(compose.uiTooling)
+        debugImplementation(libs.androidx.ui.tooling.preview)
         implementation(libs.kotlinx.datetime)
     }
