+++ multipaz-compose/src/commonMain/composeResources/drawable/presentment_icon_success.xml
