+++ multipaz-compose/src/commonMain/kotlin/org/multipaz/compose/ui/Icon.kt
@@ -1,4 +1,4 @@
-package com.android.identity.appsupport.ui
+package org.multipaz.compose.ui
 
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.AccountBalance
