+++ multipaz-compose/src/commonMain/kotlin/org/multipaz/compose/ui/Util.kt
@@ -1,4 +1,4 @@
-package com.android.identity.appsupport.ui
+package org.multipaz.compose.ui
 
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material3.MaterialTheme
