+++ multipaz-compose/src/commonMain/kotlin/org/multipaz/compose/ui/consent/ConsentModalBottomSheet.kt
@@ -1,4 +1,4 @@
-package com.android.identity.appsupport.ui.consent
+package org.multipaz.compose.ui.consent
 
 import androidx.compose.foundation.ScrollState
 import androidx.compose.foundation.background
@@ -41,27 +41,28 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.style.TextDecoration
 import androidx.compose.ui.unit.dp
-import com.android.identity.appsupport.ui.getOutlinedImageVector
+import com.android.identity.appsupport.ui.consent.ConsentDocument
+import org.multipaz.compose.ui.getOutlinedImageVector
 import com.android.identity.request.Claim
 import com.android.identity.request.MdocClaim
 import com.android.identity.request.Request
 import com.android.identity.trustmanagement.TrustPoint
-import identitycredential.identity_appsupport.generated.resources.Res
-import identitycredential.identity_appsupport.generated.resources.consent_modal_bottom_sheet_button_cancel
-import identitycredential.identity_appsupport.generated.resources.consent_modal_bottom_sheet_button_more
-import identitycredential.identity_appsupport.generated.resources.consent_modal_bottom_sheet_button_share
-import identitycredential.identity_appsupport.generated.resources.consent_modal_bottom_sheet_card_art_description
-import identitycredential.identity_appsupport.generated.resources.consent_modal_bottom_sheet_data_element_icon_description
-import identitycredential.identity_appsupport.generated.resources.consent_modal_bottom_sheet_headline_share_with_known_requester
-import identitycredential.identity_appsupport.generated.resources.consent_modal_bottom_sheet_headline_share_with_unknown_requester
-import identitycredential.identity_appsupport.generated.resources.consent_modal_bottom_sheet_share_and_stored_by_known_requester
-import identitycredential.identity_appsupport.generated.resources.consent_modal_bottom_sheet_share_and_stored_by_unknown_requester
-import identitycredential.identity_appsupport.generated.resources.consent_modal_bottom_sheet_share_with_known_requester
-import identitycredential.identity_appsupport.generated.resources.consent_modal_bottom_sheet_share_with_unknown_requester
-import identitycredential.identity_appsupport.generated.resources.consent_modal_bottom_sheet_verifier_icon_description
-import identitycredential.identity_appsupport.generated.resources.consent_modal_bottom_sheet_wallet_privacy_policy
-import identitycredential.identity_appsupport.generated.resources.consent_modal_bottom_sheet_warning_icon_description
-import identitycredential.identity_appsupport.generated.resources.consent_modal_bottom_sheet_warning_verifier_not_in_trust_list
+import identitycredential.multipaz_compose.generated.resources.Res
+import identitycredential.multipaz_compose.generated.resources.consent_modal_bottom_sheet_button_cancel
+import identitycredential.multipaz_compose.generated.resources.consent_modal_bottom_sheet_button_more
+import identitycredential.multipaz_compose.generated.resources.consent_modal_bottom_sheet_button_share
+import identitycredential.multipaz_compose.generated.resources.consent_modal_bottom_sheet_card_art_description
+import identitycredential.multipaz_compose.generated.resources.consent_modal_bottom_sheet_data_element_icon_description
+import identitycredential.multipaz_compose.generated.resources.consent_modal_bottom_sheet_headline_share_with_known_requester
+import identitycredential.multipaz_compose.generated.resources.consent_modal_bottom_sheet_headline_share_with_unknown_requester
+import identitycredential.multipaz_compose.generated.resources.consent_modal_bottom_sheet_share_and_stored_by_known_requester
+import identitycredential.multipaz_compose.generated.resources.consent_modal_bottom_sheet_share_and_stored_by_unknown_requester
+import identitycredential.multipaz_compose.generated.resources.consent_modal_bottom_sheet_share_with_known_requester
+import identitycredential.multipaz_compose.generated.resources.consent_modal_bottom_sheet_share_with_unknown_requester
+import identitycredential.multipaz_compose.generated.resources.consent_modal_bottom_sheet_verifier_icon_description
+import identitycredential.multipaz_compose.generated.resources.consent_modal_bottom_sheet_wallet_privacy_policy
+import identitycredential.multipaz_compose.generated.resources.consent_modal_bottom_sheet_warning_icon_description
+import identitycredential.multipaz_compose.generated.resources.consent_modal_bottom_sheet_warning_verifier_not_in_trust_list
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
 import org.jetbrains.compose.resources.ExperimentalResourceApi
