+++ multipaz-compose/src/commonMain/kotlin/org/multipaz/compose/ui/passphrase/PassphraseEntryField.kt
@@ -1,4 +1,4 @@
-package com.android.identity.appsupport.ui.passphrase
+package org.multipaz.compose.ui
 
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Arrangement
@@ -15,12 +15,13 @@ import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.focus.FocusRequester
 import androidx.compose.ui.focus.focusRequester
@@ -34,35 +35,44 @@ import androidx.compose.ui.text.input.TransformedText
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import com.android.identity.securearea.PassphraseConstraints
+import identitycredential.multipaz_compose.generated.resources.Res
+import identitycredential.multipaz_compose.generated.resources.passphrase_entry_field_passphrase_is_weak
+import identitycredential.multipaz_compose.generated.resources.passphrase_entry_field_pin_is_weak
 import kotlinx.coroutines.delay
-import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
+import kotlinx.coroutines.runBlocking
+import org.jetbrains.compose.resources.getString
 
-// U+2022 Bullet, see https://en.wikipedia.org/wiki/Bullet_(typography)
-//
-private const val BULLET = "\u2022"
 
+/**
+ * A composable for entering a passphrase or PIN.
+ *
+ * Three parameters are passed to the [onChanged] callback, the first is the current passphrase,
+ * the second is whether this meets requirements, and they third is a boolean which is set to
+ * `true` only if the user pressed the "Done" button on the virtual keyboard.
+ *
+ * Note that [onChanged] will be fired right after the composable is first shown. This is to
+ * enable the calling code to update e.g. a "Next" button based on whether the currently
+ * entered passphrase meets requirements.
+ *
+ * @param constraints the constraints for the passphrase.
+ * @param checkWeakPassphrase if true, checks and disallows for weak passphrase/PINs and also
+ *   shows a hint if one is input
+ * @param onChanged called when the user is entering text or pressing the "Done" IME action
+ */
 @Composable
-internal fun PassphrasePromptInputField(
+fun PassphraseEntryField(
     constraints: PassphraseConstraints,
-    showKeyboard: StateFlow<Boolean>,
-    onChanged: suspend (passphrase: String, donePressed: Boolean) -> Boolean,
+    checkWeakPassphrase: Boolean = false,
+    onChanged: (passphrase: String, meetsRequirements: Boolean, donePressed: Boolean) -> Unit,
 ) {
-    val coroutineScope = rememberCoroutineScope()
-    var inputText by remember { mutableStateOf("") }
-
-    val imeAction = if (constraints.isFixedLength()) {
-        ImeAction.None
-    } else {
-        if (inputText.length >= constraints.minLength) {
-            ImeAction.Done
-        } else {
-            ImeAction.None
-        }
-    }
-
+    // if no imeAction specified define the default to be ImeAction.Done
     val focusRequester = remember { FocusRequester() }
 
+    var inputText by remember { mutableStateOf("") }
+    var passphraseAnalysis by remember {
+        mutableStateOf(PassphraseAnalysis(false, null))
+    }
     var obfuscateAll by remember { mutableStateOf(false) }
     val scope = rememberCoroutineScope()
     // Put boxes around the entered chars for fixed length and six or fewer characters
@@ -78,12 +88,12 @@ internal fun PassphrasePromptInputField(
                     val digit =
                         if (digitIndex == inputText.length - 1) {
                             if (obfuscateAll) {
-                                BULLET
+                                "\u2022"  // U+2022 Bullet
                             } else {
                                 inputText[digitIndex].toString()
                             }
                         } else if (digitIndex < inputText.length) {
-                            BULLET
+                            "\u2022"  // U+2022 Bullet
                         } else {
                             ""
                         }
@@ -96,8 +106,8 @@ internal fun PassphrasePromptInputField(
                         textAlign = TextAlign.Center,
                         style = MaterialTheme.typography.headlineLarge.copy(
                             color = MaterialTheme.colorScheme.primary,
-                            // make the BULLET be a bit bolder/bigger
-                            fontWeight = if (digit == BULLET) {
+                            // make the "dot" \u2022 be a bit bolder/bigger
+                            fontWeight = if (digit == "\u2022") {
                                 FontWeight.Black
                             } else {
                                 MaterialTheme.typography.headlineLarge.fontWeight
@@ -109,6 +119,7 @@ internal fun PassphrasePromptInputField(
             }
         }
     }
+
     Row(
         modifier = Modifier.fillMaxWidth(),
         horizontalArrangement = Arrangement.Center,
@@ -118,16 +129,18 @@ internal fun PassphrasePromptInputField(
             modifier = Modifier
                 .padding(8.dp)
                 .focusRequester(focusRequester),
-            onValueChange = { newText ->
-                if (newText.length > constraints.maxLength) {
+            onValueChange = {
+                if (it.length > constraints.maxLength) {
                     return@BasicTextField
                 }
                 if (constraints.requireNumerical) {
-                    if (!(newText.all { it.isDigit() })) {
+                    if (!(it.all { it.isDigit() })) {
                         return@BasicTextField
                     }
                 }
-                inputText = newText
+
+                inputText = it
+                passphraseAnalysis = analyzePassphrase(inputText, constraints, checkWeakPassphrase)
 
                 obfuscateAll = false
                 scope.launch {
@@ -138,40 +151,27 @@ internal fun PassphrasePromptInputField(
                     inputText = value
                 }
 
-                coroutineScope.launch {
-                    if (onChanged(inputText, false)) {
-                        inputText = ""
-                    }
-                }
+                onChanged(inputText, passphraseAnalysis.meetsRequirements, false)
             },
             singleLine = true,
             textStyle = MaterialTheme.typography.headlineMedium.copy(color = MaterialTheme.colorScheme.primary),
             decorationBox = decorationBox,
             keyboardOptions = KeyboardOptions(
-                keyboardType = if (constraints.requireNumerical) {
-                    KeyboardType.NumberPassword
-                } else {
-                    KeyboardType.Password
-                },
-                imeAction = imeAction,
+                keyboardType = if (constraints.requireNumerical) KeyboardType.NumberPassword else KeyboardType.Password,
+                imeAction = ImeAction.Done,
             ),
             keyboardActions = KeyboardActions(
                 onDone = {
-                    if (inputText.length >= constraints.minLength) {
-                        coroutineScope.launch {
-                            if (onChanged(inputText, true)) {
-                                inputText = ""
-                            }
-                        }
-                    }
+                    onChanged(inputText, passphraseAnalysis.meetsRequirements, true)
                 }
             ),
             visualTransformation = { text ->
+                val mask = '\u2022'
                 val result = if (text.isNotEmpty()) {
                     if (obfuscateAll) {
-                        BULLET.repeat(text.text.length)
+                        mask.toString().repeat(text.text.length)
                     } else {
-                        BULLET.repeat(text.text.length - 1) + text.last()
+                        mask.toString().repeat(text.text.length - 1) + text.last()
                     }
                 } else {
                     ""
@@ -192,7 +192,88 @@ internal fun PassphrasePromptInputField(
         )
     }
 
-    if (showKeyboard.collectAsState().value == true) {
-        focusRequester.requestFocus()
+    passphraseAnalysis.weakHint?.let {
+        Row(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(8.dp),
+            horizontalArrangement = Arrangement.Center,
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            Text(
+                text = it,
+                style = MaterialTheme.typography.bodyMedium,
+                color = MaterialTheme.colorScheme.error
+            )
+        }
     }
+
+    LaunchedEffect(Unit) {
+        // Bring up keyboard when entering screen
+        focusRequester.requestFocus()
+        passphraseAnalysis = analyzePassphrase(inputText, constraints, checkWeakPassphrase)
+        // Fire initially so caller can adjust e.g. sensitivity of a possible "Next" button
+        onChanged(inputText, passphraseAnalysis.meetsRequirements, false)
+    }
+}
+
+private data class PassphraseAnalysis(
+    val meetsRequirements: Boolean,
+    val weakHint: String?,
+)
+
+private fun analyzePassphrase(
+    passphrase: String,
+    constraints: PassphraseConstraints,
+    checkWeakPassphrase: Boolean
+): PassphraseAnalysis {
+    // For a fixed-length passphrase, never give any hints until user has typed it in.
+    if (constraints.isFixedLength()) {
+        if (passphrase.length < constraints.minLength) {
+            return PassphraseAnalysis(meetsRequirements = false, weakHint = null)
+        }
+    }
+
+    if (passphrase.length < constraints.minLength) {
+        return PassphraseAnalysis(meetsRequirements = false, weakHint = null)
+    }
+
+    if (checkWeakPassphrase && isWeakPassphrase(passphrase)) {
+        return PassphraseAnalysis(
+            meetsRequirements = false,
+            weakHint = if (constraints.requireNumerical)
+                runBlocking { getString(Res.string.passphrase_entry_field_pin_is_weak) }
+            else
+                runBlocking { getString(Res.string.passphrase_entry_field_passphrase_is_weak) },
+        )
+    }
+
+    return PassphraseAnalysis(
+        meetsRequirements = true,
+        weakHint = null
+    )
+}
+
+private fun isWeakPassphrase(passphrase: String): Boolean {
+    if (passphrase.isEmpty()) {
+        return true
+    }
+
+    // Check all characters being the same
+    if (passphrase.all { it.equals(passphrase.first()) }) {
+        return true
+    }
+
+    // Check consecutive characters (e.g. 1234 or abcd)
+    var nextChar = passphrase.first().inc()
+    for (n in IntRange(1, passphrase.length - 1)) {
+        if (passphrase[n] != nextChar) {
+            return false
+        }
+        nextChar = nextChar.inc()
+    }
+
+    // TODO: add more checks
+
+    return true
 }
