+++ samples/testapp/src/commonMain/kotlin/com/android/identity/testapp/ui/PresentmentScreen.kt
@@ -2,17 +2,15 @@ package com.android.identity.testapp.ui
 
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.runtime.Composable
-import com.android.identity.appsupport.ui.presentment.Presentment
 import com.android.identity.appsupport.ui.presentment.PresentmentModel
 import com.android.identity.testapp.App
-import com.android.identity.testapp.TestAppSettingsModel
 import com.android.identity.testapp.TestAppPresentmentSource
-import com.android.identity.testapp.TestAppUtils
 import identitycredential.samples.testapp.generated.resources.Res
 import identitycredential.samples.testapp.generated.resources.app_icon
 import identitycredential.samples.testapp.generated.resources.app_name
 import org.jetbrains.compose.resources.painterResource
 import org.jetbrains.compose.resources.stringResource
+import org.multipaz.compose.ui.presentment.Presentment
 
 private const val TAG = "PresentmentScreen"
