+++ samples/testapp/src/commonMain/kotlin/com/android/identity/testapp/ui/QrCodesScreen.kt
@@ -9,8 +9,8 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.android.identity.appsupport.ui.qrcode.ShowQrCodeDialog
-import com.android.identity.appsupport.ui.qrcode.ScanQrCodeDialog
+import org.multipaz.compose.ui.qrcode.ShowQrCodeDialog
+import org.multipaz.compose.ui.qrcode.ScanQrCodeDialog
 
 @Composable
 fun QrCodesScreen(
