+++ settings.gradle.kts
@@ -54,6 +54,7 @@ include(":identity-android-legacy")
 include(":identityctl")
 include(":mrtd-reader")
 include(":mrtd-reader-android")
+include(":multipaz-compose")
 include(":jpeg2k")
 include(":samples:testapp")
 include(":samples:preconsent-mdl")
